Use stream with formdata
Since the previous code reads all the file content to memory,
it breaks `axios-logger` because it stringifies.

Instead, this commit uses stream data so that the logger won't
write the file content. Also, it reduces memory usage as well.

Additionally, we set unlimited file size to avoid the client side failure.
If the file size is too large, the server side will reject.