Align `lib` with `autify-cli` in `tsconfig.json`
This package is mainly used by `autify-cli` but there are a slight
differences between them and we couldn't capture an issue only
happens on `autify-cli`'s build.

This commit adds `lib` property to align with `autify-cli` so that
we can capture future regressions on this repository's build instead.

Along with it, we update GitHub Actions not to fail fast to validate
all environments regardless of any early failures.